Move shared code to a new crate `winit-common`
