+++ .github/workflows/ci.yml
@@ -184,7 +184,18 @@ jobs:
 
     - name: Test winit Android
       if: contains(matrix.platform.target, 'android')
-      run: cargo $CMD test -p winit-android --target=${{ matrix.platform.target }} --features native-activity --no-run
+      run: cargo $CMD test -p winit-android --features native-activity --no-run
+
+    - name: Test winit Common (EventHandler)
+      run: cargo $CMD test -p winit-common --features event-handler --no-run
+
+    - name: Test winit Common (CF)
+      if: contains(matrix.platform.target, 'apple')
+      run: cargo $CMD test -p winit-common --features core-foundation --no-run
+
+    - name: Test winit Common (XKB)
+      if: contains(matrix.platform.target, 'linux')
+      run: cargo $CMD test -p winit-common --features xkb,x11,wayland --no-run
 
     - name: Test winit Orbital
       if: contains(matrix.platform.target, 'redox')
