+++ Cargo.toml
@@ -13,6 +13,7 @@ rust-version = "1.80"
 # `winit` has no version here to allow using it in dev deps for docs.
 winit = { path = "." }
 winit-android = { version = "0.0.0", path = "winit-android" }
+winit-common = { version = "0.0.0", path = "winit-common" }
 winit-core = { version = "0.0.0", path = "winit-core" }
 winit-orbital = { version = "0.0.0", path = "winit-orbital" }
 winit-win32 = { version = "0.0.0", path = "winit-win32" }
@@ -172,12 +173,13 @@ wayland = [
     "sctk",
     "ahash",
     "memmap2",
+    "winit-common/wayland",
 ]
 wayland-csd-adwaita = ["sctk-adwaita", "sctk-adwaita/ab_glyph"]
 wayland-csd-adwaita-crossfont = ["sctk-adwaita", "sctk-adwaita/crossfont"]
 wayland-csd-adwaita-notitle = ["sctk-adwaita"]
 wayland-dlopen = ["wayland-backend/dlopen"]
-x11 = ["x11-dl", "bytemuck", "percent-encoding", "xkbcommon-dl/x11", "x11rb"]
+x11 = ["x11-dl", "bytemuck", "percent-encoding", "xkbcommon-dl/x11", "x11rb", "winit-common/x11"]
 
 [build-dependencies]
 cfg_aliases.workspace = true
@@ -208,6 +210,7 @@ winit-android.workspace = true
 block2.workspace = true
 dispatch2.workspace = true
 objc2.workspace = true
+winit-common = { workspace = true, features = ["core-foundation", "event-handler"] }
 
 # AppKit
 [target.'cfg(target_os = "macos")'.dependencies]
@@ -358,6 +361,7 @@ wayland-backend = { workspace = true, optional = true }
 wayland-client = { workspace = true, optional = true }
 wayland-protocols = { workspace = true, optional = true }
 wayland-protocols-plasma = { workspace = true, optional = true }
+winit-common = { workspace = true, features = ["xkb"] }
 x11-dl = { workspace = true, optional = true }
 x11rb = { workspace = true, optional = true, features = [
     "allow-unsafe-code",
