+++ src/platform_impl/apple/mod.rs
@@ -1,10 +1,5 @@
-//! Apple/Darwin-specific implementations
-
 #[cfg(target_os = "macos")]
 mod appkit;
-mod event_handler;
-mod event_loop_proxy;
-mod notification_center;
 #[cfg(not(target_os = "macos"))]
 mod uikit;
