+++ src/platform_impl/apple/uikit/event_loop.rs
@@ -26,8 +26,8 @@ use winit_core::event_loop::{
 use winit_core::monitor::MonitorHandle as CoreMonitorHandle;
 use winit_core::window::{Theme, Window as CoreWindow};
 
-use super::super::notification_center::create_observer;
 use super::app_state::{send_occluded_event_for_all_windows, AppState};
+use super::notification_center::create_observer;
 use super::{app_state, monitor, MonitorHandle};
 use crate::platform_impl::Window;
