+++ src/platform_impl/apple/uikit/notification_center.rs
@@ -1,0 +1,1 @@
+../appkit/notification_center.rs
