--- src/platform_impl/linux/common/mod.rs
@@ -1,1 +1,0 @@
-pub mod xkb;
