+++ winit-common/src/core_foundation/mod.rs
@@ -1,0 +1,3 @@
+mod event_loop_proxy;
+
+pub use event_loop_proxy::*;
