+++ winit-common/src/lib.rs
@@ -1,0 +1,8 @@
+//! Winit implementation helpers.
+
+#[cfg(feature = "core-foundation")]
+pub mod core_foundation;
+#[cfg(feature = "event-handler")]
+pub mod event_handler;
+#[cfg(feature = "xkb")]
+pub mod xkb;
